Producto Form Completo. Ya guarda Cantidad Ya guarda Precios Costo y Venta Inventario de Inicio de un Producto Inventario de Dia de un Producto

+++ xfinance-app-backend/src/main/java/com/gigti/xfinance/backend/data/ProductoVenta.java
@@ -20,7 +20,7 @@ public class ProductoVenta extends AbstractEntity{
     @JoinColumn
     private Producto producto;
 
-    private Integer quantitySell;
+    private double quantitySell;
 
     @Temporal(TemporalType.TIMESTAMP)
     private Date sellDate;
